requirements.txt
# StayWise AI — pip install -r requirements.txt

# Step 1 — Data Pipeline
pandas>=2.0.0
numpy>=1.24.0
PyYAML>=6.0
python-dotenv>=1.0.0

# Step 2 — RAG
sentence-transformers>=2.2.2
chromadb>=0.4.0

# Step 3 — LLM (Groq - Free)
groq>=0.9.0

# Step 5 — API
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
